课程搜索模块

+++ xc-framework-model/src/main/java/com/xuecheng/framework/domain/course/CoursePub.java
@@ -32,8 +32,8 @@ public class CoursePub implements Serializable {
     private String charge;
     private String valid;
     private String qq;
-    private Float price;
-    private Float price_old;
+    private Double price;
+    private Double price_old;
     private String expires;
     private String teachplan;//课程计划
     @Column(name="pub_time")
